feat: add TriangleDistanceToOriginComparator

+++ src/main/java/ch/hslu/ad/sw08/Point.java
@@ -1,6 +1,6 @@
 package ch.hslu.ad.sw08;
 
-public record Point(int x, int y) {
+public record Point(double x, double y) {
     public double distanceTo(final Point otherPoint) {
         return Math.hypot(
             Math.abs(otherPoint.x - x),
